% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_sleep.R
\name{plot_sleep_over_time}
\alias{plot_sleep_over_time}
\title{A function to plot sleep over time.}
\usage{
plot_sleep_over_time(Person)
}
\arguments{
\item{Person}{The user's data}
}
\value{
A ggplot2 object, prints to screen
}
\description{
Returns a line plot plotting sleep over time.  Includes sleep
duration and time asleep (in hours).
}
\examples{
plot_sleep_weekday(EX)

}
