% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_sleep.R
\name{plot_sleep_restless_prop}
\alias{plot_sleep_restless_prop}
\title{A function to plot the proportion of restless sleep over time.}
\usage{
plot_sleep_restless_prop(Person)
}
\arguments{
\item{Person}{The user's data}
}
\value{
A ggplot2 object, prints to screen
}
\description{
Returns a line plot plotting the proportion of restless sleep 
over time.  The proportion is calculated as the difference between sleep
duration and time spent asleep over sleep duration.
}
\examples{
plot_sleep_restless_prop(EX)

}
