% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_sleep.R
\name{tidy_sleep_over_time}
\alias{tidy_sleep_over_time}
\title{A function to make the sleep data tidy.}
\usage{
tidy_sleep_over_time(Person)
}
\arguments{
\item{Person}{The user's data}
}
\value{
A data frame with the columns date, sleep_time, and mins
}
\description{
Returns a line plot plotting sleep over time.  Includes sleep
duration and time asleep (in hours).
}
\examples{
plot_sleep_over_time(EX)

}
